refactor(sw): create sw folder for service-worker related files
