+++ src/sw/register.js
